Set the memory cap to 4GB for hadd step 1.5 and beyond in 1l1tau,0l2tau
The input files to hadd steps 1.5 and beyond are massive in 0l+2tau and 1l+1tau files, which causes these steps to consume more than 2GB of memory. We could either run dedicated jobs for each subcategory, create the subcategories after hadd step 1 or just increase the memory required by these jobs. The first option would generate too many jobs when running with full systematics, so this is a no-go. The second option would require some time to implement and is subject to human errors. This leaves only the third option as the viable one for the short term.

4GB (= 2 nodes worth of memory) should be sufficient enough. We could make it more efficient by asking memory dynamically based on the input file sizes if it turns out to be a bottleneck in the workflow.