Disable early object-level preselection in all analyses
Because of a bug fixed in 02ad2236ec94dbacd72f19bba7d40f987d9bfd5f, the number of tight electrons and muons counted in the Ntuple post-production step is overestimating the actual number of tight electrons and muons. The effect is that we're probably rejecting more events in low lepton multiplicity channels due to signal region veto. The bug also affects lepton FR, jet->tau FR and charge flip analyses. However, the actual impact has to be very small, because the lepton was incorrectly considered as tight only if its reco-pt > 5/7 GeV but cone-pt < 10 GeV. The problem never even surfaced in the sync exercise.