Multiply HH reweighting weight with the unweighted number of HH events
The reweighting normalized the event yields to 1, regardless of the sample size. This has to be corrected for by multiplying the event weight returned by HH framework with the unweighted number of events, or equivalently with the number of entries in HH denominator histogram.
Related: https://github.com/HEP-KBFI/tth-htt/issues/82
Compiles, otherwise untested.