Do not consider PS variation if weights are available but event sums not
In HH analyses, the single top samples of 2017 era have PS weights, but the samples are not post-processed recently and are therefore lacking corresponding event sums for each PS variation that are used to compute the lumiscale factors. Thus, the most straightforward way to tackle this is to disable the PS variation for single-top samples (and for other samples that are in similar situation).
https://github.com/HEP-KBFI/tth-htt/issues/127