Do not read & write CSVv2 b-tag score of assoc jet in 2018
Bug introduced while solving https://github.com/HEP-KBFI/tth-htt/issues/136
During the last Ntuple production of 2018, the input Ntuples were not required to have the presence of this branch. Now that none of the branches we read are "optional" (ie they are allowed to be missing from the input Ntuples), the next post-processing of 2018 Ntuples would fail.