Set hadronic taus correctly in data/MC correction interface
The data/MC correction interface had a single function that allowed to set 1-4 hadronic taus all at once, by setting the SF parameters (such as pT, gen PDG ID and eta) to some default values. The function then checked given arguments against those placeholder values when deciding on the number of hadronic taus.
However, it all falls (and fell) apart when these setter functions are misused. For instance, in 2l+2tau channel also the decay mode of the hadronic tau was passed to the interface which differed from the placedholder value. This in effect caused problems because the arguments that "spilled over" made up "third" hadronic tau with bogus parameters.
The solution is to explicate the interface: exactly N x 3 parameters for each setter that is supposed to take N hadronic taus as input.