Remove manual sorting of leading electrons in charge flip measurement
This is now done automatically after shifting the electron pT, given the systematics. Compiles, otherwise untested. The data/MC corrections are now derived from the electron collection the energy scale of which is shifted based on the choice of systematics (previously it was just the nominal electron collection). Related: https://github.com/HEP-KBFI/hh-multilepton/issues/18