Record event sums found with both conventions of gen weight application
Also, do not enable HTXS for the sync exercise. Apparently, I had produced a sync Ntuple with the HTXS before, but never copied it to /hdfs and updated samples dictionaries accordinly. Will have to recreate the Ntuples at some point, so that we could test the HTXS functionality using only the sync Ntuple.