Bugfix: read all matching gen photons in RecoLeptonReader
Our Ntuples store reconstructed particles in one branch and matchin generator-level particles in other branches. For instance, muons are in the 'Muons' branch while the photons that are matched to the muon are in the 'Muon_genPhoton' branch. Even if the reconstructed particle has no counterpart at the generator level, we store a "dummy" particle that has all fileds (4-momentum, status, flags) set to 0. This ensures that the list of reconstructed and generator level particles are always consistent, even if there is no match.
In recent development I "expanded" the GenPhoton collection to include non-final state photons. By default, the GenPhotonReader is configured to read only the final state particles because that's the normal use case (outside of gen photon filter). However, this introduced a regeression where the RecoMuonReader and RecoElectronReader try to read generator level photons, but get fewer than expected in return. The solution is to not limit the gen photons to final state when reading them in the RecoLeptonReader class.

https://github.com/HEP-KBFI/hh-multilepton/issues/36