Allow multiple inputs to MC closure and fakes in addSystFakeRates
In bbWW analysis our plan is to derive the non-closure systematics from comparing the fake contribution in the SR to the MC closure regions with
- fully hadronic ttbar events in SL channel; and
- semileptonic ttbar events in DL channel.
The most straightforward way to accomplish this is to use hadd stage 1 inputs instead of hadd stage 1.5 or stage2 files. The caveat is that there are two SL ttbar samples in every era, meaning that we'd need to specify multiple hadd stage 1 files as input. Of course, we could still get away by using the sample that has the highest number of events but there's no good reason to cut corners like that.

The changes don't change the default behavior of estimating the non-closure. Tested.