Exclude samples w/o the LHE scale weights from the scale variation
Problem surfaced in HH->bbWW SL resonant analysis, where the nominal template of spin-2 900 GeV sample was 2x larger than both sides of the LHE envelope. The issue was caused by the fact that this resonant mass point was covered by two samples, where one sample had the LHE scale weights while the other doesn't. If the LHE scale weights are missing in a sample, then the sample dictionary instead stores the nominal event counts of that particular sample. This is wrong because the LHE envelope is determined from the events that have the LHE scale weights (so one sample), while the lumiscale factor that normalizes the samples for a given variation was based on two samples. The solution is to reset the event counts that correspond to LHE scale variation to 0 if the sample is missing the necessary LHE scale weights. Tested.