Possibility to obtain shapes for every PDF member and other systematics
Saving histograms for each of the 33 PDF members adds an extra 120MB to
the memory footprint. Given that our heaviest analysis jobs just need
about 1GB of memory, then those extra MB won't blow up our memory
budget. For more recent PDF sets (such as NNPDF3.x), there are 103 PDF
members so even in that case the required memory to run a job increases
about 350MB, well below the 2GB our analysis jobs have.

The only catch is that the hadd jobs may die if we want to run the
analysis with full systematics and also have the histograms for each PDF
member in one go. Still, it may be worthwhile to try this out.